feat: add business logic layer for characters and add separate API
Added charactersSlice, and at the same time prepared an API file so
that I can already write full-fledged business logic. The
charactersSlice already has Thunk, which is processed using
extraReducers. In addition, I passed the store through the Provider,
thereby checking the functionality of the written logic. By dispatching
Thunk an array of 10 objects is returned in state.characters.characters