feat: display data in Characters.js
Now we can observe data about the first ten characters, later we will
see the rest with the help of pagination and also to make it more
convenient, we completely process the code in Characters.js using the
status, depending on which we see either loading, or data, or an error,
which is directly displayed to the user using jsx.

P.S. In extraBuilders where we process case fetchCharacters.fulfilled,
payload.results, instead of just payload, need to be concated to state.characters.