style: add styling for the character component
Now we have the opportunity to see in a beautiful way the name, date of birth,
weight and avatar of the characters, again, so far only on the first page.
Also, to get the avatars of the characters, I had to use another resource.
Therefore, in the application we can observe changes in the Character.jsx
file on line 25, now we use our own utility to get the character id,
which we throw further into the Character component, with the help of
this id and the link, a picture of one or another character is displayed