feat: upgrade pagination
I upgraded the pagination, now we have the opportunity, reaching the
last page in the portion when clicked, to see the next one and the same
way in reverse order. I also replaced the data about left-right numbers
instead of a variables with hooks useState.

Despite the upgrade, the code clearly needs refactoring in the near
future.