refactor: refactor code in multiple files
I refactored Character.jsx, Paginator.jsx & charactersSlice.jsx. In the
last commit, I mentioned the need to refactor the paginator, now the
code in it has become cleaner and more readable. The changes in the rest
of the files were due to the addition of a status check to useEffect. In
principle, this does not affect anything, but before that the status
never returned to the 'idle' state, that is, in the end it was
constantly 'succeeded'. In addition, we dispatched thunk without knowing
what our status was, which could lead to undesirable consequences.