feat: add ui for character profile
Now we have the ability to view character information on click. I don’t
particularly like the logic itself, because if the api provided the
character by default with an id, we could only do a search by id,
otherwise we have to make a separate request, which wastes resources.
So, in short, we use useParam to find out the id in the link, after
which we dispatch thunk to which we pass id and as a response we get an
object with information about the character.