fix: fix non-aborted thunk when a new one is dispatched
Let's imagine that we click on several pages and we have bad internet,
and so with each click we send a request, which is not canceled when the
page is changed. All this will spill out into the fact that after a
couple of clicks, after waiting for a certain page to load, we will
first get characters from the last pages we clicked.

Therefore, I came to the conclusion that in useEffect I get a promise
that I then abort. This may not be the correct solution, but for now
this is what came to my mind. Can be changed later during refactoring.