fix: fix cancellation of request
Now when we scroll through the pages, the loading status does not
disappear if the thunk was aborted with abort. Also, now we use a custom
request method, while, as before, we cancel the old request if a new one
was made.

To understand the changes, you should look at thunk fetchCharacters in
the last commit and in this one. We can observe that using the custom
method, we pass the config to it as the second parameter, which we then
also insert as the second parameter into the get request. All this
allows us to cancel the request in thunk.