style: add styling for planets section
I styled the planet section in a similar way to the characters.
Made a couple of changes to the character styles, and also added class
variables.

Notes: I strongly advise not to refer to the current styles, because they are
not well thought out. Basically, they are made for a more pleasant work
with application.