feat: add root component
A root component has been added, where all sections are reflected. There
is no bll here, although there is such an opportunity through a request.
At least I don't see the need for it right now.

Also, a couple of changes were made to the general styles and a simple
handling of the 404 page was added (without a separate component).