feat: add ui for species section
The Species.jsx and SpecieCard.jsx component has been added. The first
contains the main logic for displaying SpecieCard.jsx for each object.
In simple words, the logic is repeated with the previous ones,
Planets.jsx, Characters.jsx.

By the way, the code with the calculation of pageSize in planets,
characters & species Slices was commented out. To avoid the situation
when the length of the array of objects changes on the last page and
therefore the pageSize changes, and this leads to the rendering of
non-existent pages using the paginator.