feat: add user interface for specie profile
In addition to displaying the profile, a few console.log and other
unnecessary material have also been removed, while this does not affect
the change in the application. UI of the profile is the same as with the
previous components such as PlanetProfile.jsx and CharacterProfile.jsx.

P.S. The selectors in specieSlice.jsx were also changed, namely the
status and error selectors, because their paths were not correct.