refactor: remove duplicate profile components
Now there is only one component for displaying the profile of an object
- Profile.jsx. At the same time, the ProfileExcerpt component was also
added, in which the logic for which Excerpt should be shown, was added.
Also, now there is no need to pass an object of two parameters to the
dispatch loadProfile. Since after refactoring we use pathname, which
we pass further to the API request.