fix: fix extra re-rendering and paginator
In Entities.jsx, the useState hook content has been added instead of the
usual variable. In addition, a conditional operation with a status has
been added to useEffect.

As for the paginator, a check was added there that if the current page
is more than the number of pages, paginate the first page.