refactor: refactor multiple files, see description...
Firstly, I changed the conditional check in Header.jsx where it checked if
the link is active to add this style. Instead of 'yes?' && do this then,
now: '....' ? 'do this' : 'nothing then'. To avoid unnecessary errors and
notifications in the console.

Secondly, I changed the date display format in FilmExcerpt.jsx.
From 1977-05-25 to Wed May 25 1977 (example).