style: add parallax effect
The title was supposed to have a parallax effect, which is why another
intro section was added which contains the bg and the logo. Scrolling
through, we get to the main section, where the previous work is already
located. Everything is accompanied by a parallax effect.

It is important to note that all the styles have not yet been worked out
to the end, so the work looks quite raw.