feat: add preloader
Instead of the usual paragraph with text, the preloader has been added
as a separate component so that it can be easily used in different
places.

In addition, in Profile.jsx, let content has been replaced with the
useState() hook & the status handling has also been moved under the
useEffect() hook to avoid unnecessary re-renders.